jsoned/keywords/min_properties_keyword.py
from jsoned.errors.schema_parse_error import SchemaParseError
from jsoned.json_core import AssertionKeyword, JsonSchema
from jsoned.types import JsonObject, JsonType
from jsoned.validators.core_validators import CompoundValidator
from jsoned.validators.object_validators import MinimumPropertiesValidator


class MinimumPropertiesKeyword(AssertionKeyword):
    key = "minProperties"

    def apply(self, document: JsonSchema, node: JsonObject, validator: CompoundValidator):
        if node[self.key].type != JsonType.NUMBER:
            raise SchemaParseError.for_invalid_keyword_value(node, self.key, JsonType.NUMBER)

        validator[self.key] = MinimumPropertiesValidator(expected_minimum=int(node[self.key]))
